move logs
